refactor: move ToolsConfiguration to exporter-common and eliminate magic string duplication
The three default tool extension property name strings ("io.camunda.tool:name",
"inbound.type", "io.camunda.tool:") were duplicated across three places:
Tools.java, ExporterConfiguration.ToolsConfiguration (camunda-exporter), and
ExporterConfiguration.ToolsConfiguration (rdbms-exporter).

Introduce io.camunda.zeebe.exporter.common.tools.ToolsConfiguration as the
single canonical home for both the class and its DEFAULT_* constants. All three
modules already depend on exporter-common (following the same pattern as
AuditLogConfiguration and HistoryDeletionConfiguration), so:

- Remove the duplicate nested ToolsConfiguration classes from both exporter
  ExporterConfiguration classes and replace with the shared type
- Update Tools.java to delegate its DEFAULT_* constants to the shared ones so
  existing references remain valid
- Update all handler/applier imports accordingly (7 main sources + 4 tests)
